feat: Software Templates (#13)
* docs: Add TechDocs  instructions

* feat(wip): Add templates.md and update mkdocs.yml and app-config.production.yaml

* fix(ci): Fixing linter errors

* fix: Update owner in template.yaml

* fix: Update template.yaml and add catalog-info.yaml

* fix: exclude templates from tflint

* fix: exclude templates from yaml linting, since they aren't yaml

* feat: Add Terraform control module files

Add the skel and templates to generate a Terraform Repo using sofware templating.  I'm adding the files to the Container image, but at some point we should stop doing that, since it'll make the image even larger than it already is

* fix: Add checks.yaml workflow and remove pr-checks.yaml workflow

* fix: Add Terraform control module template and update working directory in GitHub workflow

* ci(fix): Add .trivyignore file to exclude certain configurations

* fix: Add GitHub Team field to pr-catalog template and fixed the repoUrl function

* fix: Update template.yaml files with allowedOwners and githubTeam fields

* fix: Add default value for PagerDuty integration

* fix: Test Comment

* fix: Remove static_analysis job from checks.yaml

* fix: fixing some linting errors

* fix: Update catalog-info.yaml files

Trying to make the yaml file pass linting.. this as as close as I can get.

Also updated the description typos

* feat(wip): Add new files and directories for Terraform with Terragrunt template

* feat(wip): Add Terraform and Terragrunt configuration files***

* fix: Update GitHub workflow versions

* fix(typo): Update terraform-docs version in checks.yaml

* fix: Update minReplicas value to 1 for dev in hpa.yaml

The Database Dev uses is a bit small, and results in errors during the CI deployment since the DB starts getting to many connections. eventually k8s reconciles  this but the check gets a red x when it shouldn't

* fix: Add google-beta provider to terragrunt template