chore: Change targets for Domain Templates to use relative paths. (#52)
* chore: Change targets for Domain Templates to use relative paths.

Add test.yaml for managing Cloud Platforms

This commit adds a new file `test.yaml` under `backstage/templates/BITS/domains` directory. The `test.yaml` file defines a `Domain` resource with metadata and specifications for managing Cloud Platforms. This change is made to support the management of Cloud Platforms in the application.

Co-authored-by: dependabot[bot] <support@github.com>

* chore: Remove test.yaml from BITS domains

The test.yaml file in the BITS domains directory is no longer needed and has been removed. This change removes the test.yaml file from the repository.

* chore: Update targets for Domain Templates to use relative paths

This commit updates the targets for Domain Templates in the BITS systems.yaml file to use relative paths instead of absolute URLs. The modified paths are changed from:
- https://github.com/broadinstitute/backstage/tree/main/backstage/templates/BITS/systems/kubernetes.yaml
- https://github.com/broadinstitute/backstage/tree/main/backstage/templates/BITS/systems/hpc-comm.yaml
- https://github.com/broadinstitute/backstage/tree/main/backstage/templates/BITS/systems/hpc-drag.yaml
- https://github.com/broadinstitute/backstage/tree/main/backstage/templates/BITS/systems/seqprod.yaml

to:
- ./systems/kubernetes.yaml
- ./systems/hpc-comm.yaml
- ./systems/hpc-drag.yaml
- ./systems/seqprod.yaml

---------

Co-authored-by: dependabot[bot] <support@github.com>