feat: Migrate to new backend system
The new backend system simplified the backstage configuration, in particular plugins.
The impact of this is a cleaner app, that easier for us to manage and reason about.

More info can be found here https://backstage.io/docs/backend-system/building-backends/migrating/

Mostly working, except the user can't login, saving here so I can
troubleshoot and have a good place to roll back to