feat: Add Slack for Soundcheck
Adding slack
notifications to
soundcheck

In order to get Slack notifications from your Slack app, it must be added to a Slack channel. The Slack channel can be either a public or a private channel that lives in the same workspace that you selected when creating the Slack app.

Right-click on the desired channel in Slack, select View Channel Details, click on the Integrations tab, and then click on Add Apps. Search for your Slack app by name, find it in the search results, and then click on the Add button.

https://backstage.spotify.com/docs/plugins/soundcheck/setup-and-installation#adding-the-slack-app-to-a-slack-channel