feat: add cost object as a field to create namespace template
The kube create namespace template now takes cost object as an input
which sets that label, so that it can be uses in BQ to charge back
resource ussage to the owners

also fixes a small bug in the PD plugin
older versions of this plugin sometimes generate this error, and bumping
the versions fixes this issue

```
Unauthorized
You don't have the required permissions to
perform this action. See README for more details.
```