chore(templates): Curate "recommended" templates
I want to have a single link to all our templates that we want to
advertize to non-BITS people. Alaric suggested using "recommended" for
this purpose instead of creating a new "external" flag. I like this
implementation because it allows us to re-use the existing "Recommended"
section that groups all of these templates at the top of the page even
without a filter.